Creación de los endpoints para manejar las imagenes y audios
Creación de los endpoints para manejar las imagenes y audios
