+++ src/controllers/login-controller.js
@@ -5,7 +5,7 @@ const {responseServer, responseGeneral} = require('../helpers/response-result');
 
 const postLogin = async (request, response) => {
     const {username, password} = request.body;
-    await Accounts.findOne({username: username}, {_id:1, status:1, password:1, role:1, name:1, lastname:1, URLPhoto:1})
+    await Accounts.findOne({username: username}, {_id:1, status:1, password:1, role:1, name:1, lastname:1, URL:1})
     .then(async (account) =>{  
         if (account) {
             if(account.status == 1){
